requirements.txt
django
django-heroku
gunicorn
psycopg2
pytz
whitenoise
django-crispy-forms
